Refactor font builders
This removes the duplication between the font builders, and fixes issues with image resolution - the image defined in the JSON will now be looked up in the same folder as defined in the project.
