+++ src/Types/TR1/Misc/TR1FontBuilder.cs
@@ -1,83 +1,17 @@
-﻿using System.Drawing;
-using TRImageControl;
-using TRImageControl.Packing;
+﻿using TRImageControl.Packing;
 using TRLevelControl.Helpers;
 using TRLevelControl.Model;
-using TRXInjectionTool.Control;
-using TRXInjectionTool.Util;
 
 namespace TRXInjectionTool.Types.TR1.Misc;
 
-public class TR1FontBuilder : InjectionBuilder
+public class TR1FontBuilder : FontBuilder
 {
-    private readonly List<GlyphDef> _glyphDefs;
-    private readonly Dictionary<string, TRImage> _imageCache;
-
-    public override string ID => "font";
-
     public TR1FontBuilder()
-    {
-        _glyphDefs = DeserializeFile<List<GlyphDef>>("Resources/TR1/Font/glyph_info.json");
-        _glyphDefs.Sort((g1, g2) => g1.mesh_num.CompareTo(g2.mesh_num));
-        _imageCache = new();
-    }
+        : base(TRGameVersion.TR1) { }
 
-    public TRImage GetImage(GlyphDef glyph)
-    {
-        string path = Path.Combine("Resources/TR1/Font", glyph.filename);
-        if (!_imageCache.ContainsKey(path))
-        {
-            _imageCache[path] = new(path);
-        }
-
-        TRImage image = _imageCache[path];
-        if (!_imageCache.ContainsKey(glyph.filename))
-        {
-            _imageCache[glyph.filename] = new(glyph.filename);
-        }
-
-        Rectangle bounds = new(glyph.x, glyph.y, glyph.w, glyph.h);
-        return image.Export(bounds);
-    }
-
-    public override List<InjectionData> Build()
+    protected override TRLevelBase Pack(TRSpriteSequence font, List<TRTextileRegion> regions)
     {
         TR1Level caves = _control1.Read($"Resources/{TR1LevelNames.CAVES}");
-
-        TRSpriteSequence font = new();
-        List<TRTextileRegion> regions = new();
-
-        foreach (GlyphDef glyph in _glyphDefs)
-        {
-            TRSpriteTexture texture = new()
-            {
-                Bounds = new(glyph.x, glyph.y, glyph.w, glyph.h),
-                Alignment = new()
-                {
-                    Left = glyph.l,
-                    Top = glyph.t,
-                    Right = glyph.r,
-                    Bottom = glyph.b,
-                },
-            };
-
-            font.Textures.Add(texture);
-
-            regions.Add(new()
-            {
-                Bounds = texture.Bounds,
-                Image = GetImage(glyph),
-                Segments = new()
-                {
-                    new()
-                    {
-                        Index = glyph.mesh_num,
-                        Texture = texture,
-                    },
-                },
-            });
-        }
-
         ResetLevel(caves, 1);
         TR1TexturePacker packer = new(caves);
         packer.AddRectangles(regions);
@@ -86,8 +20,6 @@ public class TR1FontBuilder : InjectionBuilder
         caves.Sprites[TR1Type.FontGraphics_S_H] = font;
 
         _control1.Write(caves, MakeOutputPath(TRGameVersion.TR1, $"Debug/{ID}.phd"));
-
-        InjectionData data = InjectionData.Create(caves, InjectionType.General, ID);
-        return new() { data };
+        return caves;
     }
 }
